# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Events-API is service for storing Kubernetes and Containerum events in MongoDB.",
    "title": "Events-API",
    "version": "0.0.1"
  },
  "paths": {
    "/all": {
      "get": {
        "tags": [
          "AllEvents"
        ],
        "summary": "Get all events.",
        "operationId": "AllResourcesChangesEvents",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "ws",
            "in": "query"
          },
          {
            "type": "string",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "websocket response",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
...
    "/namespaces/{namespace}/all": {
      "get": {
        "tags": [
          "AllEvents"
        ],
        "summary": "Get all events in namespace.",
        "operationId": "AllNamespaceResourcesChangesEvents",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ws",
            "in": "query"
          },
          {
            "type": "string",
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "websocket response",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/namespaces/{namespace}/selected": {
      "get": {
        "tags": [
          "AllEvents"
        ],
        "summary": "Get selected events in namespace.",
        "operationId": "SelectedNamespaceResourcesChangesEvents",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "namespace",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "name": "ws",
            "in": "query"
          },
          {
            "type": "string",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "name": "res",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "websocket response",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    },
    "/selected": {
      "get": {
        "tags": [
          "AllEvents"
        ],
        "summary": "Get selected events.",
        "operationId": "SelectedResourcesChangesEvents",
        "parameters": [
          {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string",
            "name": "X-User-Role",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "format": "uuid",
            "name": "X-User-ID",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "name": "ws",
            "in": "query"
          },
          {
            "type": "string",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "name": "res",
            "in": "query"
          }
        ],
        "responses": {
          "101": {
            "description": "websocket response",
            "schema": {
              "$ref": "#/definitions/EventsList"
            }
          },
          "default": {
            "description": "cherry error",
            "schema": {
              "$ref": "#/definitions/allResourcesChangesEventsDefaultBody"
            }
          }
        },
        "x-method-visibility": "public"
      }
    }
  },
  "definitions": {
...
    "Event": {
      "description": "Event -- Containerum event",
      "type": "object",
      "properties": {
        "details": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "event_kind": {
          "$ref": "#/definitions/EventKind"
        },
        "event_name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "event_time": {
          "type": "string",
          "x-go-name": "Time"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "resource_name": {
          "type": "string",
          "x-go-name": "ResourceName"
        },
        "resource_namespace": {
          "type": "string",
          "x-go-name": "ResourceNamespace"
        },
        "resource_type": {
          "$ref": "#/definitions/ResourceType"
        },
        "resource_uid": {
          "type": "string",
          "x-go-name": "ResourceUID"
        }
      },
      "x-go-package": "github.com/containerum/events-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "EventKind": {
      "type": "string",
      "x-go-package": "github.com/containerum/events-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "EventsList": {
      "type": "object",
      "properties": {
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Event"
          },
          "x-go-name": "Events"
        }
      },
      "x-go-package": "github.com/containerum/events-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "ResourceType": {
      "type": "string",
      "x-go-package": "github.com/containerum/events-api/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "allResourcesChangesEventsDefaultBody": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "description": "Fields -- represents additional key-value fields of error",
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-package": "github.com/containerum/cherry"
        },
        "id": {
          "description": "ErrID -- represents unique error ID",
          "type": "object",
          "properties": {
            "kind": {
              "description": "ErrKind -- represents kind of error",
              "type": "integer",
              "format": "uint64",
              "x-go-package": "github.com/containerum/cherry"
            },
            "sid": {
              "description": "ErrSID -- represents service ID of error",
              "type": "string",
              "x-go-package": "github.com/containerum/cherry"
            }
          },
          "x-go-package": "github.com/containerum/cherry"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-gen-location": "operations",
      "x-go-package": "github.com/containerum/cherry"
    }
  },
...
}